+++ client/src/Page/Master/Designation/NewDesignation.js
@@ -25,6 +25,9 @@ import ViewDesignation from "./ViewDesignation";
 const NewDesignation = () => {
   const dispatch = useDispatch();
   const designationData = useSelector(selectdesignationData);
+  const [successMessage, setSuccessMessage] = useState("");
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState(null);
 
   const [formData, setFormData] = useState({
     DesignationName: "",
@@ -53,21 +56,52 @@ const NewDesignation = () => {
     event.preventDefault();
     try {
       await dispatch(createDesignationData(formData));
+
+      // Handle success response
+      setSuccessMessage("Designation Created successfully!");
+
+       // Clear the success message after 2 seconds
+       setTimeout(() => {
+        setSuccessMessage("");
+      }, 2000);
+
       setFormData({ DesignationName: "" });
       dispatch(fetchDesignationData());
       handleClose();
     } catch (error) {
-      console.error("Error creating department:", error);
+      console.error("Error creating designation:", error);
+     setError("Error adding designation");
+    }
+    finally {
+      setLoading(false);
     }
   };
 
   const handleDeleteDesignation = async (DesignationID) => {
     try {
+      const confirmed = window.confirm(
+        "Are you sure you want to delete this Designation?"
+      );
+      if (!confirmed) {
+        setLoading(false);
+        return; // Exit function if user cancels deletion
+      }
+
       await dispatch(deleteDesignationData(DesignationID));
-      console.log(DesignationID)
+
+      setSuccessMessage("Designation deleted successfully!");
+
+      setTimeout(() => {
+        setSuccessMessage("");
+      }, 2000);
+      // console.log(DesignationID)
       dispatch(fetchDesignationData());
     } catch (error) {
       console.error("Error deleting designation:", error);
+     setError("Error deleting designation");
+    }
+    finally {
+      setLoading(false);
     }
   };
 
@@ -76,6 +110,25 @@ const NewDesignation = () => {
     
    
         <div style={{ padding: "10px", border: "2px solid #dddddd" }}>
+        <div
+            style={{
+              backgroundColor: successMessage
+                ? "green"
+                : error
+                ? "red"
+                : "transparent",
+                color: successMessage
+                ? "white"
+                : error
+                ? "white"
+                : "transparent",
+                padding:"2px 10px",
+            }}
+          >
+            {loading && <p style={{margin:"5px"}}>Loading...</p>}
+            {error && <p style={{margin:"5px"}}>{error}</p>}
+            {successMessage && <p style={{margin:"0px"}}>{successMessage}</p>}
+          </div>
           <div
             style={{
               display: "flex",
